fixed code duplication
